Renamed some actions
Change the action name from "integrate_between_layers" to
"compute_depth_integral" in the code and update the corresponding
function names and parser configurations accordingly.
